+++ build-config/fragments/dev.js
@@ -5,31 +5,41 @@
 
 const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const RemoveEmptyScriptsPlugin = require('webpack-remove-empty-scripts');
 
 const contentBase = path.resolve(__dirname, '..', '..');
 
 module.exports = {
     mode: 'development',
     devServer: {
-        contentBase: [contentBase],
-        publicPath: 'localhost:8080/dist/',
-        watchContentBase: true,
-        watchOptions: {
-            ignored: [
-                /.build_cache/,
-                /.chrome/,
-                /.git/,
-                /node_modules/,
-                /bower_components/,
-                /coverage/,
-                /build-config/,
-                /test/
-            ]
-        }
+        port: 8080,
+        static: [
+            {
+                directory: contentBase,
+                staticOptions: {},
+                publicPath: '/',
+                serveIndex: true,
+                watch: {
+                    ignored: [
+                        /.build_cache/,
+                        /.chrome/,
+                        /docs/,
+                        /node_modules/,
+                        /bower_components/,
+                        /coverage/,
+                        /build-config/,
+                        /test/,
+                        /vendor/
+                    ]
+                }
+            }
+        ]
     },
     plugins: [
+        new RemoveEmptyScriptsPlugin(),
         new MiniCssExtractPlugin({
-            filename: 'css/[name].css'
+            filename: 'css/videojs.wavesurfer.css',
+            chunkFilename: 'css/[id].css'
         })
     ]
 };